bug fixes, new flag
